+++ camelon-website/camelon-front/src/components/Display.js
@@ -2,8 +2,8 @@ import React from "react";
 import Overview from "./Modes/Overview";
 import Pinmap from "./Modes/Pinmap";
 import Heatmap from "./Modes/Heatmap";
-import Aboutus from "./Modes/aboutus";
-import ChoroplethMap from "./Modes/ChoroplethMap"
+import "bootstrap/dist/css/bootstrap.min.css";
+import ChoroplethMap from "./Modes/ChoroplethMap";
 
 export default function Display(props) {
   if (props.data === "overview") {
@@ -12,9 +12,7 @@ export default function Display(props) {
     return <Pinmap />;
   } else if (props.data === "heat_map") {
     return <Heatmap />;
-  } else if (props.data === "aboutus") {
-    return <Aboutus />;
   } else if (props.data === "choropleth_map") {
-    return <ChoroplethMap />
+    return <ChoroplethMap />;
   }
 }
