+++ camelon-website/camelon-front/src/components/Modes/chart/BarChart.js
