+++ camelon-website/camelon-front/src/components/Modes/emermode/BatteryAssault.js
@@ -1,7 +1,6 @@
-import React from "react";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
-import EmergencyBanner from "../cards/emerContact";
+import EmergencyBanner from "../../cards/emerContact";
 const EmerContact = [
   {
     id: "1",
@@ -9,18 +8,10 @@ const EmerContact = [
     callNum: "191",
     image:
       "https://logo-th.com/wp-content/uploads/2018/12/%e0%b8%aa%e0%b8%b3%e0%b8%99%e0%b8%b1%e0%b8%81%e0%b8%87%e0%b8%b2%e0%b8%99%e0%b8%95%e0%b8%b3%e0%b8%a3%e0%b8%a7%e0%b8%88%e0%b9%81%e0%b8%ab%e0%b9%88%e0%b8%87%e0%b8%8a%e0%b8%b2%e0%b8%95%e0%b8%b4.jpg",
-  },
-  {
-    id: "2",
-    name: "หน่วยกู้ชีพวชิรพยาบาล",
-    callNum: "1554",
-    image:
-      "https://logo-th.com/wp-content/uploads/2018/12/%e0%b8%aa%e0%b8%b3%e0%b8%99%e0%b8%b1%e0%b8%81%e0%b8%87%e0%b8%b2%e0%b8%99%e0%b8%95%e0%b8%b3%e0%b8%a3%e0%b8%a7%e0%b8%88%e0%b9%81%e0%b8%ab%e0%b9%88%e0%b8%87%e0%b8%8a%e0%b8%b2%e0%b8%95%e0%b8%b4.jpg",
-  },
-];
-function Murder() {
-  return (
-    <main>
+  },];
+function BatteryAssault() {
+    return (
+      <main>
       <Row>
         {EmerContact.map((data) => (
           <Col sm={4} key={data.id}>
@@ -33,7 +24,7 @@ function Murder() {
         ))}
       </Row>
     </main>
-  );
-}
-
-export default Murder;
+    );
+  }
+  
+  export default BatteryAssault;
