+++ camelon-website/camelon-front/src/components/Modes/emermode/Gambling.js
@@ -1,5 +1,5 @@
 import React from "react";
-import EmergencyBanner from "../cards/emerContact";
+import EmergencyBanner from "../../cards/emerContact";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
 const EmerContact = [
