+++ camelon-website/camelon-front/src/components/Modes/emermode/Murder.js
@@ -1,6 +1,7 @@
+import React from "react";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
-import EmergencyBanner from "../cards/emerContact";
+import EmergencyBanner from "../../cards/emerContact";
 const EmerContact = [
   {
     id: "1",
@@ -11,15 +12,15 @@ const EmerContact = [
   },
   {
     id: "2",
-    name: "ศุนย์คุ้มครองและช่วยเหลือนักเรียนนักศึกษาซึ่งถูกล่วงละเมิดทางเพศ",
-    callNum: "1579",
+    name: "หน่วยกู้ชีพวชิรพยาบาล",
+    callNum: "1554",
     image:
       "https://logo-th.com/wp-content/uploads/2018/12/%e0%b8%aa%e0%b8%b3%e0%b8%99%e0%b8%b1%e0%b8%81%e0%b8%87%e0%b8%b2%e0%b8%99%e0%b8%95%e0%b8%b3%e0%b8%a3%e0%b8%a7%e0%b8%88%e0%b9%81%e0%b8%ab%e0%b9%88%e0%b8%87%e0%b8%8a%e0%b8%b2%e0%b8%95%e0%b8%b4.jpg",
   },
 ];
-function SexualAbuse() {
-    return (
-      <main>
+function Murder() {
+  return (
+    <main>
       <Row>
         {EmerContact.map((data) => (
           <Col sm={4} key={data.id}>
@@ -32,7 +33,7 @@ function SexualAbuse() {
         ))}
       </Row>
     </main>
-    );
-  }
-  
-  export default SexualAbuse;
+  );
+}
+
+export default Murder;
