+++ camelon-website/camelon-front/src/components/Modes/emermode/SexualAbuse.js
@@ -1,6 +1,6 @@
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
-import EmergencyBanner from "../cards/emerContact";
+import EmergencyBanner from "../../cards/emerContact";
 const EmerContact = [
   {
     id: "1",
@@ -11,13 +11,13 @@ const EmerContact = [
   },
   {
     id: "2",
-    name: "รับแจ้งรถหาย ถูกขโมย",
-    callNum: "1192",
+    name: "ศุนย์คุ้มครองและช่วยเหลือนักเรียนนักศึกษาซึ่งถูกล่วงละเมิดทางเพศ",
+    callNum: "1579",
     image:
       "https://logo-th.com/wp-content/uploads/2018/12/%e0%b8%aa%e0%b8%b3%e0%b8%99%e0%b8%b1%e0%b8%81%e0%b8%87%e0%b8%b2%e0%b8%99%e0%b8%95%e0%b8%b3%e0%b8%a3%e0%b8%a7%e0%b8%88%e0%b9%81%e0%b8%ab%e0%b9%88%e0%b8%87%e0%b8%8a%e0%b8%b2%e0%b8%95%e0%b8%b4.jpg",
   },
 ];
-function TheftBurglary() {
+function SexualAbuse() {
     return (
       <main>
       <Row>
@@ -35,4 +35,4 @@ function TheftBurglary() {
     );
   }
   
-  export default TheftBurglary;
+  export default SexualAbuse;
