+++ camelon-website/camelon-front/src/components/Modes/emermode/TheftBurglary.js
@@ -1,6 +1,6 @@
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
-import EmergencyBanner from "../cards/emerContact";
+import EmergencyBanner from "../../cards/emerContact";
 const EmerContact = [
   {
     id: "1",
@@ -8,8 +8,16 @@ const EmerContact = [
     callNum: "191",
     image:
       "https://logo-th.com/wp-content/uploads/2018/12/%e0%b8%aa%e0%b8%b3%e0%b8%99%e0%b8%b1%e0%b8%81%e0%b8%87%e0%b8%b2%e0%b8%99%e0%b8%95%e0%b8%b3%e0%b8%a3%e0%b8%a7%e0%b8%88%e0%b9%81%e0%b8%ab%e0%b9%88%e0%b8%87%e0%b8%8a%e0%b8%b2%e0%b8%95%e0%b8%b4.jpg",
-  },];
-function BatteryAssault() {
+  },
+  {
+    id: "2",
+    name: "รับแจ้งรถหาย ถูกขโมย",
+    callNum: "1192",
+    image:
+      "https://logo-th.com/wp-content/uploads/2018/12/%e0%b8%aa%e0%b8%b3%e0%b8%99%e0%b8%b1%e0%b8%81%e0%b8%87%e0%b8%b2%e0%b8%99%e0%b8%95%e0%b8%b3%e0%b8%a3%e0%b8%a7%e0%b8%88%e0%b9%81%e0%b8%ab%e0%b9%88%e0%b8%87%e0%b8%8a%e0%b8%b2%e0%b8%95%e0%b8%b4.jpg",
+  },
+];
+function TheftBurglary() {
     return (
       <main>
       <Row>
@@ -27,4 +35,4 @@ function BatteryAssault() {
     );
   }
   
-  export default BatteryAssault;
+  export default TheftBurglary;
