+++ camelon-website/camelon-front/src/components/controller/emerController.js
@@ -2,10 +2,10 @@ import React, { useState } from "react";
 import Container from "react-bootstrap/esm/Container";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
-import Heading from "./Heading";
+import Heading from "../Heading";
 import Card from "react-bootstrap/Card";
 
-import NavbarEmer from "./navbarEmer";
+import NavbarEmer from "../navbarEmer";
 
 export default function EmerController() {
   const [state, setState] = useState({});
