+++ camelon-website/camelon-front/src/page/About.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Sidebar2 from "./Sidebar2";
+import Sidebar2 from "../components/Sidebar2";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
 import Container from "react-bootstrap/esm/Container";
